Add `isPaid` property to track order payment status
Introduced a new `isPaid` property in the `Order` model to track
whether an order has been paid. Updated the `OrderRequest` and
`OrderResponse` DTOs to include this property. Modified the
`OrderController` to set `isPaid` to `false` by default when
creating new orders.

Enhanced the `OrderService` to handle updates to the `isPaid`
property. Updated the `OrderMapper` to map the `isPaid` property
between the model and DTOs.

Added a new migration `20251117062047_AddIsPayToOrder` to add
the `isPaid` column to the `Order` table in the database schema.
Updated the EF Core model snapshot to reflect this change.